split Trace SDK in sub-namespaces and create new area for contrib
move all tests in the same root directory - will be easy to follow

remove errant test and whitelist the library code for coverage instead of tests

inherit JaegerExporter from Zipkin since we are using the zipkin protocol to export to Jaeger
