+++ examples/JaegerExporterExample.php
@@ -3,11 +3,11 @@
 declare(strict_types=1);
 require __DIR__ . '/../vendor/autoload.php';
 
-use OpenTelemetry\Sdk\Trace\AlwaysOnSampler;
+use OpenTelemetry\Contrib\Jaeger\Exporter as JaegerExpoter;
 use OpenTelemetry\Sdk\Trace\Attributes;
 use OpenTelemetry\Sdk\Trace\Clock;
-use OpenTelemetry\Sdk\Trace\JaegerExporter;
-use OpenTelemetry\Sdk\Trace\SimpleSpanProcessor;
+use OpenTelemetry\Sdk\Trace\Sampler\AlwaysOnSampler;
+use OpenTelemetry\Sdk\Trace\SpanProcessor\SimpleSpanProcessor;
 use OpenTelemetry\Sdk\Trace\TracerProvider;
 use OpenTelemetry\Trace as API;
 
@@ -20,7 +20,7 @@ $samplingResult = $sampler->shouldSample(
     API\SpanKind::KIND_INTERNAL
 );
 
-$exporter = new JaegerExporter(
+$exporter = new JaegerExpoter(
     'jaegerExporterExample',
     'http://jaeger:9412/api/v2/spans'
 );
