+++ tests/Sdk/Integration/AlwaysOffSamplerTest.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenTelemetry\Tests\Sdk\Integration;
+
+use OpenTelemetry\Sdk\Trace\Sampler\AlwaysOffSampler;
+use OpenTelemetry\Sdk\Trace\SamplingResult;
+use OpenTelemetry\Trace as API;
+use PHPUnit\Framework\TestCase;
+
+class AlwaysOffSamplerTest extends TestCase
+{
+    public function testAlwaysOffSampler()
+    {
+        $sampler = new AlwaysOffSampler();
+        $decision = $sampler->shouldSample(
+            null,
+            '4bf92f3577b34da6a3ce929d0e0e4736',
+            '00f067aa0ba902b7',
+            'test.opentelemetry.io',
+            API\SpanKind::KIND_INTERNAL
+        );
+        $this->assertEquals(SamplingResult::NOT_RECORD, $decision->getDecision());
+    }
+
+    public function testAlwaysOnSamplerDescription()
+    {
+        $sampler = new AlwaysOffSampler();
+        $this->assertEquals('AlwaysOffSampler', $sampler->getDescription());
+    }
+}
