+++ tests/Sdk/Integration/AlwaysParentSamplerTest.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenTelemetry\Tests\Sdk\Integration;
+
+use OpenTelemetry\Sdk\Trace\Sampler\AlwaysParentSampler;
+use OpenTelemetry\Sdk\Trace\SamplingResult;
+use OpenTelemetry\Sdk\Trace\SpanContext;
+use OpenTelemetry\Trace as API;
+use PHPUnit\Framework\TestCase;
+
+class AlwaysParentSamplerTest extends TestCase
+{
+    public function testRecordAlwaysParentSamplerDecision()
+    {
+        $parentContext = new SpanContext(
+            '4bf92f3577b34da6a3ce929d0e0e4736',
+            '00f067aa0ba902b7',
+            0x1
+        );
+        $sampler = new AlwaysParentSampler();
+        $decision = $sampler->shouldSample(
+            $parentContext,
+            '4bf92f3577b34da6a3ce929d0e0e4736',
+            '00f067aa0ba902b7',
+            'test.opentelemetry.io',
+            API\SpanKind::KIND_INTERNAL
+        );
+        $this->assertEquals(SamplingResult::RECORD_AND_SAMPLED, $decision->getDecision());
+    }
+
+    public function testSkipAlwaysParentSamplerDecision()
+    {
+        $parentContext = new SpanContext(
+            '4bf92f3577b34da6a3ce929d0e0e4736',
+            '00f067aa0ba902b7',
+            0
+        );
+        $sampler = new AlwaysParentSampler();
+        $decision = $sampler->shouldSample(
+            $parentContext,
+            '4bf92f3577b34da6a3ce929d0e0e4736',
+            '00f067aa0ba902b7',
+            'test.opentelemetry.io',
+            API\SpanKind::KIND_INTERNAL
+        );
+        $this->assertEquals(SamplingResult::NOT_RECORD, $decision->getDecision());
+    }
+
+    public function testNullAlwaysParentSamplerDecision()
+    {
+        $sampler = new AlwaysParentSampler();
+        $decision = $sampler->shouldSample(
+            null,
+            '4bf92f3577b34da6a3ce929d0e0e4736',
+            '00f067aa0ba902b7',
+            'test.opentelemetry.io',
+            API\SpanKind::KIND_INTERNAL
+        );
+        $this->assertEquals(SamplingResult::NOT_RECORD, $decision->getDecision());
+    }
+
+    public function testAlwaysParentSamplerDescription()
+    {
+        $sampler = new AlwaysParentSampler();
+        $this->assertEquals('AlwaysParentSampler', $sampler->getDescription());
+    }
+}
