+++ tests/Sdk/Unit/Trace/SpanProcessor/SimpleSpanProcessorTest.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenTelemetry\Tests\Sdk\Unit\Trace\SpanProcessor;
+
+use OpenTelemetry\Sdk\Trace\Exporter;
+use OpenTelemetry\Sdk\Trace\Span;
+use OpenTelemetry\Sdk\Trace\SpanProcessor\SimpleSpanProcessor;
+use PHPUnit\Framework\TestCase;
+
+class SimpleSpanProcessorTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function shouldCallExporterOnEnd()
+    {
+        $exporter = self::createMock(Exporter::class);
+        $exporter->expects($this->atLeastOnce())->method('export');
+
+        (new SimpleSpanProcessor($exporter))->onEnd(
+            self::createMock(Span::class)
+        );
+    }
+}
