+++ tests/Sdk/Unit/Trace/TracerTest.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenTelemetry\Tests\Sdk\Unit\Trace;
+
+use OpenTelemetry\Sdk\Trace\SpanProcessor;
+use OpenTelemetry\Sdk\Trace\Tracer;
+use PHPUnit\Framework\TestCase;
+
+class TracerTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function spanProcessorsShouldBeCalledWhenNewSpanIsCreated()
+    {
+        $processor = self::createMock(SpanProcessor::class);
+        $processor->expects($this->atLeastOnce())->method('onStart');
+
+        /** @var \OpenTelemetry\Sdk\Trace\SpanProcessor $processor */
+        $tracer = new Tracer([$processor]);
+
+        $tracer->startAndActivateSpan('test.span');
+    }
+
+    /**
+     * @test
+     */
+    public function spanProcessorsShouldBeCalledWhenActiveSpanIsEnded()
+    {
+        $processor = self::createMock(SpanProcessor::class);
+
+        $processor->expects($this->atLeastOnce())->method('onEnd');
+
+        /** @var \OpenTelemetry\Sdk\Trace\SpanProcessor $processor */
+        $tracer = new Tracer([$processor]);
+
+        $tracer->startAndActivateSpan('test.span');
+        $tracer->endActiveSpan();
+    }
+}
